+++ includes/class-wp-statistics-schedule.php
@@ -100,14 +100,17 @@ class Schedule
             wp_unschedule_event(wp_next_scheduled('wp_statistics_geoip_hook'), 'wp_statistics_geoip_hook');
         }
 
-        if (wp_next_scheduled('wp_statistics_geoip_hook') && 'cf' === Option::get('geoip_location_detection_method', 'maxmind')) {
+        $locationDetection = Option::get('geoip_location_detection_method', 'maxmind');
+
+        if (wp_next_scheduled('wp_statistics_geoip_hook') && 'cf' === $locationDetection) {
             wp_unschedule_event(wp_next_scheduled('wp_statistics_geoip_hook'), 'wp_statistics_geoip_hook');
         }
 
         //Construct Event
-        if ('maxmind' === Option::get('geoip_location_detection_method', 'maxmind')) {
+        if (in_array($locationDetection, ['maxmind', 'dbip'], true)) {
             add_action('wp_statistics_geoip_hook', array($this, 'geoip_event'));
         }
+
         add_action('wp_statistics_report_hook', array($this, 'send_report'));
         add_action('wp_statistics_licenses_hook', [$this, 'migrateOldLicenses']);